Pass tests for truncated histogram aggregation
